Updated remediation summary. Refactored systems modal.
