+++ src/components/RemediationDetailsTable.js
@@ -1,14 +1,10 @@
-import React, { useContext } from 'react';
+import React, { useContext, useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 
 import flatMap from 'lodash/flatMap';
 import orderBy from 'lodash/orderBy';
 
-import {
-    Button,
-    Pagination,
-    ToolbarItem, Toolbar, ToolbarContent
-} from '@patternfly/react-core';
+import { Pagination } from '@patternfly/react-core';
 
 import { sortable, TableHeader, Table, TableBody, TableVariant } from '@patternfly/react-table';
 import { RedoIcon, TimesIcon } from '@patternfly/react-icons';
@@ -20,37 +16,39 @@ import './RemediationTable.scss';
 
 import { ConnectResolutionEditButton } from '../containers/ConnectedComponents';
 import { DeleteActionsButton } from '../containers/DeleteButtons';
-import { isBeta } from '../config';
-import SystemForActionButton from './SystemForActionButton';
+//import { isBeta } from '../config';
+import { SystemForActionButton } from './SystemForActionButton';
 
 import { useFilter, usePagination, useSelector, useSorter } from '../hooks/table';
 import * as debug from '../Utilities/debug';
 
 import './RemediationDetailsTable.scss';
-import { CheckCircleIcon } from '@patternfly/react-icons';
 import { PermissionContext } from '../App';
 
 function resolutionDescriptionCell (remediation, issue) {
+    const url = buildIssueUrl(issue.id);
+
     if (issue.resolutions_available <= 1) {
-        return issue.resolution.description;
+        return (url
+            ? <React.Fragment>
+                <a href={ url }>{ issue.description }</a>
+                <br/>{ issue.resolution.description }
+            </React.Fragment>
+            : issue.resolution.description
+        );
     }
 
-    return (
-        <React.Fragment>
+    return (url
+        ? <React.Fragment>
+            <a href={ url }>{ issue.description }</a>
+            <br/>{ issue.resolution.description }&nbsp;
+            <ConnectResolutionEditButton issue={ issue } remediation={ remediation } />
+        </React.Fragment>
+        : <React.Fragment>
             { issue.resolution.description }&nbsp;
             <ConnectResolutionEditButton issue={ issue } remediation={ remediation } />
         </React.Fragment>
     );
-}
-
-function issueDescriptionCell (issue) {
-    const url = buildIssueUrl(issue.id);
-
-    if (url) {
-        return <a href={ url }>{ issue.description }</a>;
-    }
-
-    return issue.description;
 }
 
 function needsRebootCell (needsReboot) {
@@ -69,8 +67,13 @@ function needsRebootCell (needsReboot) {
     );
 }
 
-function systemsForAction(issue, remediation) {
-    return <SystemForActionButton key={ issue.id } remediation={ remediation } issue={ issue } />;
+function systemsForAction(issue, remediation, title) {
+    return (
+        <SystemForActionButton key={ issue.id }
+            remediation={ remediation }
+            issue={ issue }
+            title={ title } />
+    );
 }
 
 const SORTING_ITERATEES = [
@@ -89,20 +92,21 @@ const buildRow = (remediation) => (issue) => {
             id: issue.id,
             cells: [
                 {
-                    title: issueDescriptionCell(issue)
-                },
-                {
                     title: resolutionDescriptionCell(remediation, issue)
                 },
                 {
                     title: needsRebootCell(issue.resolution.needs_reboot)
                 },
                 {
-                    title: systemsForAction(issue, remediation)
+                    title: systemsForAction(issue, remediation, `${issue.systems.length}`)
                 },
                 {
                     title: getIssueApplication(issue),
                     props: { className: 'ins-m-nowrap' }
+                },
+                {
+                    title: systemsForAction(issue, remediation,
+                        `${remediation.resolved_count}/${issue.systems.length} remediated`) // PLACEHOLDER VALUE
                 }
             ]
         }
@@ -117,6 +121,11 @@ function RemediationDetailsTable (props) {
     const filter = useFilter();
     const selector = useSelector();
     const permission = useContext(PermissionContext);
+    const [ filterText, setFilterText ] = useState('');
+
+    useEffect(() => {
+        filter.setValue(filterText);
+    }, [ filterText ]);
 
     sorter.onChange(pagination.reset);
     filter.onChange(pagination.reset);
@@ -130,6 +139,11 @@ function RemediationDetailsTable (props) {
     selector.register(rows);
 
     const selectedIds = selector.getSelectedIds();
+
+    const activeFiltersConfig = {
+        filters: filterText.length ? [{ category: 'Action', chips: [{ name: filterText }]}] : [],
+        onDelete: () => {setFilterText(''); filter.setValue('');}
+    };
 
     return (
         <div className='test'>
@@ -167,7 +181,7 @@ function RemediationDetailsTable (props) {
                 pagination={ { ...pagination.props, itemCount: filtered.length } }
                 activeFiltersConfig={ activeFiltersConfig }
             />
-            {/* { /* <Toolbar className='ins-c-remediations-details-table__toolbar'>
+            { /* <Toolbar className='ins-c-remediations-details-table__toolbar'>
                 <ToolbarContent>
                     <ToolbarItem>
                         <SimpleTableFilter buttonTitle="" placeholder="Search actions" { ...filter.props } />
@@ -197,7 +211,7 @@ function RemediationDetailsTable (props) {
                         { ...debug.pagination }
                     />
                 </ToolbarContent>
-            </Toolbar> */}
+            </Toolbar> */ }
             {
                 rows.length > 0 ?
                     <Table
@@ -210,8 +224,6 @@ function RemediationDetailsTable (props) {
                                 title: 'Actions',
                                 transforms: [ sortable ]
                             }, {
-                                title: 'Resolution'
-                            }, {
                                 title: 'Reboot required',
                                 transforms: [ sortable ]
                             }, {
@@ -219,6 +231,9 @@ function RemediationDetailsTable (props) {
                                 transforms: [ sortable ]
                             }, {
                                 title: 'Type',
+                                transforms: [ sortable ]
+                            }, {
+                                title: 'Status',
                                 transforms: [ sortable ]
                             }]
                         }
