+++ src/components/RemediationSummary.scss
@@ -73,9 +73,9 @@
 .ins-c-remediation-summary {
     //height: 100%;
 
-    &__chart {
-        
-    }
+    // &__chart {
+    //     fill: var(--pf-global--BackgroundColor--100)
+    // }
 
     &__body {
         padding-top: var(--pf-global--spacer--lg);
